dsl: Move D and S operators to stiffness.utils
