+++ examples/seismic/stiffness/operators.py
@@ -1,7 +1,8 @@
 from devito import Eq, Operator, VectorTimeFunction, TensorTimeFunction
-from devito import div, grad, diag, solve, D, S
+from devito import div, grad, diag, solve
 from examples.seismic import PointSource, Receiver
 from examples.seismic.utils import matriz_init
+from examples.seismic.stiffness.utils import D, S
 
 def iso_elastic_tensor(model):
     def subs3D(lmbda, mu):
